+++ src/IbanNet/Validation/DefaultValidationRuleResolver.cs
@@ -10,7 +10,6 @@ namespace IbanNet.Validation
     internal class DefaultValidationRuleResolver : IValidationRuleResolver
     {
         private readonly IbanValidatorOptions _options;
-        private readonly IStructureValidationFactory _structureValidationFactory;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DefaultValidationRuleResolver" />.
@@ -18,12 +17,6 @@ namespace IbanNet.Validation
         public DefaultValidationRuleResolver(IbanValidatorOptions options)
         {
             _options = options ?? throw new ArgumentNullException(nameof(options));
-
-            _structureValidationFactory = new CachedStructureValidationFactory(
-                new CompositeStructureValidationFactory(
-                    (options.Registry ?? throw new ArgumentException(Resources.ArgumentException_Registry_is_required, nameof(options))).Providers
-                )
-            );
         }
 
         /// <inheritdoc />
@@ -38,12 +31,12 @@ namespace IbanNet.Validation
 
             if (_options.Method == ValidationMethod.Strict)
             {
-                yield return new IsMatchingStructureRule(_structureValidationFactory);
+                yield return new IsMatchingStructureRule();
             }
 
             yield return new Mod97Rule();
 
-            if (_options.Rules is null)
+            if (_options.Rules is null!)
             {
                 yield break;
             }