+++ test/IbanNet.Tests/Registry/BankStructureTests.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using FluentAssertions;
-using IbanNet.Validation;
+using IbanNet.Registry.Patterns;
 using Xunit;
 
 namespace IbanNet.Registry
@@ -8,37 +8,19 @@ namespace IbanNet.Registry
     public class BankStructureTests
     {
         [Fact]
-        public void When_creating_with_null_structure_it_should_throw()
+        public void When_creating_with_null_pattern_it_should_throw()
         {
-            string structure = null;
+            Pattern pattern = null;
 
             // Act
-            // ReSharper disable once ObjectCreationAsStatement
             // ReSharper disable once AssignNullToNotNullAttribute
-            Func<BankStructure> act = () => new BankStructure(structure, new NullStructureValidationFactory());
+            Func<BankStructure> act = () => new BankStructure(pattern);
 
             // Assert
             act.Should()
                 .Throw<ArgumentNullException>()
                 .Which.ParamName.Should()
-                .Be(nameof(structure));
-        }
-
-        [Fact]
-        public void When_creating_with_null_structure_validation_factory_it_should_throw()
-        {
-            IStructureValidationFactory structureValidationFactory = null;
-
-            // Act
-            // ReSharper disable once ObjectCreationAsStatement
-            // ReSharper disable once AssignNullToNotNullAttribute
-            Func<BankStructure> act = () => new BankStructure(string.Empty, structureValidationFactory);
-
-            // Assert
-            act.Should()
-                .Throw<ArgumentNullException>()
-                .Which.ParamName.Should()
-                .Be(nameof(structureValidationFactory));
+                .Be(nameof(pattern));
         }
     }
 }